minor fixes

+++ src/type_alias.c
@@ -1,6 +1,6 @@
 #include <type_alias.h>
-void *memcpy(void *dest, const void *src, i32 n) {
-    i32 i;
+void *memcpy(void *dest, const void *src, usize n) {
+    u32 i;
 
     for (i = 0; i < n; i++) {
         ((unsigned char *)dest)[i] = ((unsigned char *)src)[i];
@@ -9,8 +9,8 @@ void *memcpy(void *dest, const void *src, i32 n) {
     return dest;
 }
 
-void *memset(void *blk, int c, i32 n) {
-    i32 i;
+void *memset(void *blk, i32 c, usize n) {
+    u32 i;
 
     for (i = 0; i < n; i++) {
         ((unsigned char *)blk)[i] = c;
